+++ tests/TestCase/TestSuite/TestTraitTest.php
@@ -16,6 +16,7 @@ use App\AnotherExampleChildClass;
 use App\ExampleChildClass;
 use App\ExampleClass;
 use App\ExampleOfTraversable;
+use BadMethodCallException;
 use Exception;
 use PHPUnit\Framework\AssertionFailedError;
 use PHPUnit\Framework\Error\Deprecated;
@@ -27,6 +28,61 @@ use Tools\TestSuite\TestCase;
  */
 class TestTraitTest extends TestCase
 {
+    /**
+     * Test for `__call()` and `__callStatic()` magic methods
+     * @ŧest
+     */
+    public function testMagicCallAndCallStatic()
+    {
+        //Methods that use the `assertInternalType()` method
+        foreach ([
+            'assertIsArray' => ['array'],
+            'assertIsBool' => true,
+            'assertIsFloat' => 1.1,
+            'assertIsInt' => 1,
+            'assertIsObject' => new stdClass,
+            'assertIsString' => 'string',
+        ] as $assertMethod => $value) {
+            $this->{$assertMethod}($value);
+            self::{$assertMethod}($value);
+        }
+
+        //Methods that use the `assertTrue()` method, jointly to the "is" php functions
+        foreach ([
+            'assertIsCallable' => [$this, __METHOD__],
+            'assertIsIterable' => new ExampleOfTraversable,
+            'assertIsJson' => json_encode('string'),
+            'assertIsPositive' => 1,
+            'assertIsResource' => stream_context_create(),
+            'assertIsUrl' => 'http://google.com',
+        ] as $assertMethod => $value) {
+            $this->{$assertMethod}($value);
+            self::{$assertMethod}($value);
+        }
+
+        //Assertion failure
+        if (version_compare(PHP_VERSION, '7.0', '>=')) {
+            $expectedMessage = 'Failed asserting that \'string\' is of type "array".';
+        } else {
+            $expectedMessage = 'Failed asserting that false is true.';
+        }
+        $this->assertException(AssertionFailedError::class, function () {
+            $this->assertIsArray('string');
+        }, $expectedMessage);
+
+        //Missing argument
+        $this->assertException(BadMethodCallException::class, function () {
+            $this->assertIsArray();
+        }, 'Method ' . get_parent_class($this) . '::assertIsArray() expects at least 1 argument, maximum 2, 0 passed');
+
+        //Calling a no existing method or a no existing "assertIs" method
+        foreach (['assertIsNoExistingType', 'noExistingMethod'] as $method) {
+            $this->assertException(BadMethodCallException::class, function () use ($method) {
+                $this->$method('string');
+            }, 'Method ' . get_parent_class($this) . '::' . $method . '() does not exist');
+        }
+    }
+
     /**
      * Tests for `assertArrayKeysEqual()` method
      * @test
@@ -75,6 +131,7 @@ class TestTraitTest extends TestCase
             throw new Exception('right exception message');
         }, 'right exception message');
 
+        //No exception throw
         try {
             $this->assertException(Exception::class, 'time');
         } catch (AssertionFailedError $e) {
@@ -83,6 +140,20 @@ class TestTraitTest extends TestCase
             unset($e);
         }
 
+        //No existing exception or invalid exception class
+        foreach (['noExistingException', stdClass::class] as $class) {
+            try {
+                $this->assertException($class, function () {
+                    throw new Exception;
+                });
+            } catch (AssertionFailedError $e) {
+            } finally {
+                $this->assertStringStartsWith('Class `' . $class . '` does not exist or is not an Exception instance', $e->getMessage());
+                unset($e);
+            }
+        }
+
+        //Unexpected exception type
         try {
             $this->assertException(Deprecated::class, function () {
                 throw new Exception;
@@ -93,6 +164,7 @@ class TestTraitTest extends TestCase
             unset($e);
         }
 
+        //Wrong exception message
         try {
             $this->assertException(Exception::class, function () {
                 throw new Exception('Wrong');
@@ -103,6 +175,7 @@ class TestTraitTest extends TestCase
             unset($e);
         }
 
+        //Expected exception message, but no message
         try {
             $this->assertException(Exception::class, function () {
                 throw new Exception;
@@ -121,8 +194,14 @@ class TestTraitTest extends TestCase
     {
         $files = [create_tmp_file(), create_tmp_file()];
         $this->assertFileExists($files[0]);
+
+        $errorReporting = error_reporting(E_ALL & ~E_USER_DEPRECATED);
         $this->assertFileExists($files);
         $this->assertFileExists(new ExampleOfTraversable($files));
+        error_reporting($errorReporting);
+
+        $this->expectException(Deprecated::class);
+        $this->assertFileExists($files);
     }
 
     /**
@@ -133,12 +212,18 @@ class TestTraitTest extends TestCase
     {
         $this->assertFileExtension('jpg', 'file.jpg');
         $this->assertFileExtension('jpeg', 'FILE.JPEG');
+
+        $errorReporting = error_reporting(E_ALL & ~E_USER_DEPRECATED);
         $this->assertFileExtension('jpg', [
             'file.jpg',
             'file.JPG',
             'path/to/file.jpg',
             '/full/path/to/file.jpg',
         ]);
+        error_reporting($errorReporting);
+
+        $this->expectException(Deprecated::class);
+        $this->assertFileExtension('jpg', ['file.jpg', 'file.JPG']);
     }
 
     /**
@@ -149,6 +234,12 @@ class TestTraitTest extends TestCase
     {
         $files = [create_tmp_file('string'), create_tmp_file('string')];
         $this->assertFileMime($files[0], 'text/plain');
+
+        $errorReporting = error_reporting(E_ALL & ~E_USER_DEPRECATED);
+        $this->assertFileMime($files, 'text/plain');
+        error_reporting($errorReporting);
+
+        $this->expectException(Deprecated::class);
         $this->assertFileMime($files, 'text/plain');
     }
 
@@ -160,8 +251,14 @@ class TestTraitTest extends TestCase
     {
         $files = [TMP . 'noExisting1', TMP . 'noExisting2'];
         $this->assertFileNotExists($files[0]);
+
+        $errorReporting = error_reporting(E_ALL & ~E_USER_DEPRECATED);
         $this->assertFileNotExists($files);
         $this->assertFileNotExists(new ExampleOfTraversable($files));
+        error_reporting($errorReporting);
+
+        $this->expectException(Deprecated::class);
+        $this->assertFileNotExists($files);
     }
 
     /**
@@ -177,12 +274,18 @@ class TestTraitTest extends TestCase
         $this->assertFilePerms($files[0], ['0600', '0666']);
         $this->assertFilePerms($files[0], [0600, 0666]);
         $this->assertFilePerms($files[0], ['0600', 0666]);
+
+        $errorReporting = error_reporting(E_ALL & ~E_USER_DEPRECATED);
         $this->assertFilePerms($files, '0600');
         $this->assertFilePerms($files, 0600);
         $this->assertFilePerms($files, ['0600', '0666']);
         $this->assertFilePerms($files, [0600, 0666]);
         $this->assertFilePerms(new ExampleOfTraversable($files), '0600');
         $this->assertFilePerms(new ExampleOfTraversable($files), ['0600', '0666']);
+        error_reporting($errorReporting);
+
+        $this->expectException(Deprecated::class);
+        $this->assertFilePerms($files, '0600');
     }
 
     /**
@@ -197,16 +300,6 @@ class TestTraitTest extends TestCase
     }
 
     /**
-     * Tests for `assertIsArray()` method
-     * @test
-     */
-    public function testAssertIsArray()
-    {
-        $this->assertIsArray([]);
-        $this->assertIsArray(['value']);
-    }
-
-    /**
      * Tests for `assertIsArrayNotEmpty()` method
      * @test
      */
@@ -214,35 +307,17 @@ class TestTraitTest extends TestCase
     {
         $this->assertIsArrayNotEmpty(['value']);
 
-        $this->expectException(AssertionFailedError::class);
-        $this->assertIsArrayNotEmpty([]);
-    }
-
-    /**
-     * Tests for `assertIsInt()` method
-     * @test
-     */
-    public function testAssertIsInt()
-    {
-        $this->assertIsInt(1);
-    }
-
-    /**
-     * Tests for `assertIsObject()` method
-     * @test
-     */
-    public function testAssertIsObject()
-    {
-        $this->assertIsObject(new stdClass);
-    }
-
-    /**
-     * Tests for `assertIsString()` method
-     * @test
-     */
-    public function testAssertIsString()
-    {
-        $this->assertIsString('string');
+        foreach ([
+            [],
+            [[]],
+            [false],
+            [null],
+            [''],
+        ] as $array) {
+            $this->assertException(AssertionFailedError::class, function () use ($array) {
+                $this->assertIsArrayNotEmpty($array);
+            });
+        }
     }
 
     /**